:recycle: api: move files into sessions-mass-import directory
