+++ api/lib/domain/services/sessions-mass-import/sessions-import-validation-service.js
@@ -1,7 +1,7 @@
-const { UnprocessableEntityError } = require('../../application/http-errors.js');
-const { SessionWithIdAndInformationOnMassImportError, InvalidCertificationCandidate } = require('../errors.js');
-const sessionValidator = require('../validators/session-validator.js');
-const certificationCpfService = require('./certification-cpf-service.js');
+const { UnprocessableEntityError } = require('../../../application/http-errors.js');
+const { SessionWithIdAndInformationOnMassImportError, InvalidCertificationCandidate } = require('../../errors.js');
+const sessionValidator = require('../../validators/session-validator.js');
+const certificationCpfService = require('../certification-cpf-service.js');
 
 module.exports = {
   async validateSession({ session, sessionRepository, certificationCourseRepository }) {
