+++ api/lib/domain/usecases/index.js
@@ -131,7 +131,7 @@ const scorecardService = require('../../domain/services/scorecard-service.js');
 const scoringCertificationService = require('../../domain/services/scoring/scoring-certification-service.js');
 const supOrganizationParticipantRepository = require('../../infrastructure/repositories/sup-organization-participant-repository.js');
 const sessionForAttendanceSheetRepository = require('../../infrastructure/repositories/sessions/session-for-attendance-sheet-repository.js');
-const sessionsImportValidationService = require('../../domain/services/sessions-import-validation-service.js');
+const sessionsImportValidationService = require('../../domain/services/sessions-mass-import/sessions-import-validation-service.js');
 const sessionPublicationService = require('../../domain/services/session-publication-service.js');
 const sessionRepository = require('../../infrastructure/repositories/sessions/session-repository.js');
 const sessionForSupervisingRepository = require('../../infrastructure/repositories/sessions/session-for-supervising-repository.js');
