+++ api/tests/unit/domain/usecases/validate-sessions-for-mass-import_test.js
@@ -4,7 +4,7 @@ const { InvalidCertificationCandidate } = require('../../../../lib/domain/errors
 const sessionCodeService = require('../../../../lib/domain/services/session-code-service');
 const Session = require('../../../../lib/domain/models/Session');
 const certificationCpfService = require('../../../../lib/domain/services/certification-cpf-service');
-const sessionsImportValidationService = require('../../../../lib/domain/services/sessions-import-validation-service');
+const sessionsImportValidationService = require('../../../../lib/domain/services/sessions-mass-import/sessions-import-validation-service');
 const temporarySessionsStorageForMassImportService = require('../../../../lib/domain/services/sessions-mass-import/temporary-sessions-storage-for-mass-import-service');
 const { CpfBirthInformationValidation } = require('../../../../lib/domain/services/certification-cpf-service');
 const { UnprocessableEntityError } = require('../../../../lib/application/http-errors');
